+++ actions/recommendations/index.ts
@@ -1,1 +1,4 @@
 export { getMonthlyWatchlist } from "./getMonthlyWatchlist"
+export { getRecommendations } from "./getRecommendations"
+export type { Recommendations } from "./getRecommendations"
+export { getGenres } from "./getGenres"